Fixed Usage metres
Code now allows for a user to change the usage values, and the length of
the bars will change. It requires the value to have either + or - in
front, which decides whether or not to make the red bar larger or the
green bar larger. The other bar is then set to length 0.