Synced the timer image to the sheets
I synced up the time that the timer image changes to the rest of the
page, and tidied up the pictures that I had to be more similar. I also
added the text file that holds the output of the Python code to a hidden
iframe. The next decision will be whether it is better to let the
webpage or the python code worry about timing.